@@ -20,10 +20,10 @@ version = "${mod_version}"
 group = "top.theillusivec4.consecration"
 archivesBaseName = "consecration-forge"
 
-java.toolchain.languageVersion = JavaLanguageVersion.of(16)
+java.toolchain.languageVersion = JavaLanguageVersion.of(17)
 
 minecraft {
-    mappings channel: 'parchment', version: '2021.12.12-1.17.1'
+    mappings channel: 'parchment', version: '2022.03.13-1.18.2'
 
     accessTransformer = file("src/main/resources/META-INF/accesstransformer.cfg")
 
@@ -68,10 +68,6 @@ repositories {
 
 dependencies {
     minecraft "net.minecraftforge:forge:${forge_version}"
-
-    implementation fg.deobf("curse.maven:silentgear-297039:${silentgear_file}")
-
-    runtimeOnly fg.deobf("curse.maven:silentlib-242998:${silentlib_file}")
 }
 
 sourceSets {
