@@ -23,24 +23,22 @@ import javax.annotation.Nonnull;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.projectile.AbstractArrow;
 import net.minecraft.world.item.ArrowItem;
-import net.minecraft.world.item.Item;
 import net.minecraft.world.item.CreativeModeTab;
+import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
 import top.theillusivec4.consecration.common.entity.FireArrowEntity;
-import top.theillusivec4.consecration.common.registry.RegistryReference;
 
 public class FireArrowItem extends ArrowItem {
 
   public FireArrowItem() {
     super(new Item.Properties().tab(CreativeModeTab.TAB_COMBAT));
-    this.setRegistryName(RegistryReference.FIRE_ARROW);
   }
 
   @Nonnull
   @Override
   public AbstractArrow createArrow(@Nonnull Level worldIn, @Nonnull ItemStack stack,
-      @Nonnull LivingEntity shooter) {
+                                   @Nonnull LivingEntity shooter) {
     return new FireArrowEntity(worldIn, shooter);
   }
 }
