@@ -20,28 +20,27 @@
 package top.theillusivec4.consecration.common.item;
 
 import javax.annotation.Nonnull;
+import net.minecraft.world.InteractionHand;
+import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.item.Item;
 import net.minecraft.world.item.CreativeModeTab;
+import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.InteractionResult;
-import net.minecraft.world.InteractionHand;
-import top.theillusivec4.consecration.common.registry.RegistryReference;
 
 public class FireStickItem extends Item {
 
   public FireStickItem() {
     super(new Item.Properties().durability(13).tab(CreativeModeTab.TAB_COMBAT));
-    this.setRegistryName(RegistryReference.FIRE_STICK);
   }
 
   @Override
   public boolean onLeftClickEntity(ItemStack stack, Player player, Entity entity) {
 
-    if (!player.level.isClientSide && !entity.getType().fireImmune()) {
-      stack.hurtAndBreak(1, player, damager -> damager.broadcastBreakEvent(InteractionHand.MAIN_HAND));
+    if (!player.getLevel().isClientSide() && !entity.getType().fireImmune()) {
+      stack.hurtAndBreak(1, player,
+          damager -> damager.broadcastBreakEvent(InteractionHand.MAIN_HAND));
       entity.setSecondsOnFire(2);
     }
     return false;
@@ -50,9 +49,10 @@ public class FireStickItem extends Item {
   @Nonnull
   @Override
   public InteractionResult interactLivingEntity(@Nonnull ItemStack stack, Player playerIn,
-      @Nonnull LivingEntity target, @Nonnull InteractionHand hand) {
+                                                @Nonnull LivingEntity target,
+                                                @Nonnull InteractionHand hand) {
 
-    if (!playerIn.level.isClientSide && !target.getType().fireImmune()) {
+    if (!playerIn.getLevel().isClientSide() && !target.getType().fireImmune()) {
       stack.hurtAndBreak(1, playerIn, damager -> damager.broadcastBreakEvent(hand));
       target.setSecondsOnFire(2);
       return InteractionResult.SUCCESS;
